+++ projects/frontend/es-app/src/views/gamelist/IGameListView.h
@@ -18,9 +18,7 @@ public:
 
 	virtual ~IGameListView() {}
 
-	// Called when a new file is added, a file is removed, a file's metadata changes, or a file's children are sorted.
-	// NOTE: FILE_SORTED is only reported for the topmost FileData, where the sort started.
-	//       Since sorts are recursive, that FileData's children probably changed too.
+	// Called when a new file is added, a file is removed, a file's metadata changes, or when file sort changed
 	virtual void onFileChanged(FileData* file, FileChangeType change) = 0;
 	
 	// Called whenever the theme changes.